+++ src/mulle-make-compiler.sh
@@ -87,7 +87,6 @@ _compiler_debug_options()
 }
 
 
-
 # compiler is re
 compiler_sdk_parameter()
 {